#87 refactor: 챗봇 전용 헤더 컴포넌트 디렉토리 이동 및 이름 변경
